feat: random message (#2)
* deps: add rand crate

We will use this to pick a random message.

* feat: add questions.json

This contains an array of questions. We will load this and pick a
random question to send to the user.

* fix: random message not changing

This fix the issue where the message is not random in each call to
`send`. We moved the initialization to `msg` inside the loop so it get's
regenerated in each iteration.

* feat: implement random message

This fully implements the random message feature by utilizing the rand
crate's `choose` function. We load the questions from `questions.json`
to the constant QUESTIONS and parse it using `load_questions` and
finally return the result to the `choose` call.

closes #1